requirements.txt
bs4==0.0.1
Flask==0.12.2
Flask-Cors==3.0.3
html2text==2017.10.4
httplib2==0.10.3
ipaddress==1.0.19
Jinja2==2.10
JSON-log-formatter==0.2.0
mongoengine==0.15.0
flask-mongoengine
parsedatetime==2.4
python-crfsuite==0.9.5
python-dateutil==2.6.1
requests==2.18.4
simplejson==3.13.2
Werkzeug==0.13
gunicorn
gevent
pytest
cloudpickle
flask_script
blinker
h5py
PyJWT